+++ proofs/src/plonk/traces.rs
@@ -3,7 +3,7 @@
 use ff::PrimeField;
 
 use crate::{
-    plonk::{logup, permutation, trash},
+    plonk::{argument, logup, permutation},
     poly::{Coeff, LagrangeCoeff, Polynomial, commitment::PolynomialCommitmentScheme},
 };
 
@@ -16,7 +16,7 @@ pub struct ProverTrace<F: PrimeField> {
     // This field will be useful for split accumulation
     pub(crate) instance_values: Vec<Polynomial<F, LagrangeCoeff>>,
     pub(crate) lookups: Vec<logup::prover::Committed<F>>,
-    pub(crate) trashcans: Vec<trash::prover::Committed<F>>,
+    pub(crate) phase2_committed: argument::prover::Committed<F, Coeff>,
     pub(crate) permutations: permutation::prover::Committed<F>,
     pub(crate) beta: F,
     pub(crate) gamma: F,
@@ -30,7 +30,7 @@ pub struct ProverTrace<F: PrimeField> {
 pub struct VerifierTrace<F: PrimeField, PCS: PolynomialCommitmentScheme<F>> {
     pub(crate) advice_commitments: Vec<PCS::Commitment>,
     pub(crate) lookups: Vec<logup::verifier::Committed<F, PCS>>,
-    pub(crate) trashcans: Vec<trash::verifier::Committed<F, PCS>>,
+    pub(crate) phase2_committed: argument::verifier::Committed<F, PCS>,
     pub(crate) permutations: permutation::verifier::Committed<F, PCS>,
     pub(crate) beta: F,
     pub(crate) gamma: F,